Reorganized core example files and simulation manager
