+++ invertedai/helpers/simulation_manager.py
